hobby_back/requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-appconf==1.0.3
django-braces==1.13.0
django-cors-headers==3.1.1
django-filter==2.2.0
django-imagekit==4.0.2
django-rest-auth==0.9.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
oauthlib==3.1.0
pilkit==2.0
Pillow==6.2.0
psycopg2==2.8.3
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
wincertstore==0.2
